requirements.txt
python-dotenv
langchain
pinecone-client
gensim
spacy
nltk
openai-whisper
openai
scipy==1.10.1
scikit-learn
langchain-pinecone
langchain-openai
langchain-community
langsmith
langchain-cli
streamlit
ragas
jupyter
pyarrow

